requirements.txt
websocket-client
cryptography
schedule
qrcode
httpx